Support traditional and simplified characters
Previously, this was a one character toggle in the source, but there's
nothing that necessitates to make a decision at all: simply look at
both fields in the dictionary. This way, even mixed input works!

Introduce a function to avoid specifying the regex twice, which makes
the script longer than two lines, so drop that point, together with the
mention of traditional characters, from the README.